Move all views to seperate files
