+++ central_server/__init__.py
@@ -21,7 +21,11 @@ def main(global_config, **settings):
     config.add_static_view(name='static', path='static')
 
     # Configure routes
-    config.include(views)
+    config.scan()
+    config.add_route('home', '/')
+    config.add_route('cc_home', '/cc/{id}')
+    config.add_route('cc_status', '/cc/{id}/status')
+    config.add_route('cc_move', '/cc/{id}/move')
 
     # serve central_server
     return config.make_wsgi_app()
